+++ account_banking_reconciliation/__init__.py
@@ -19,6 +19,5 @@
 #    along with this program.  If not, see <http://www.gnu.org/licenses/>
 #
 ##############################################################################
-from . import account_banking_reconciliation
-from . import account_move_line
 from . import report
+from . import models
