+++ account_banking_reconciliation/models/__init__.py
@@ -1,0 +1,3 @@
+from . import account_move_line
+from . import bank_acc_rec_satetement_line
+from . import bank_acc_rec_statement
